Move ImageField to fields folder
