--- packages/tux/src/components/ImageField/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ImageField'
